# static/openapi.json
{
  "components": {
    "schemas": {
      "protobufAny": {
        "additionalProperties": {},
        "properties": {
          "@type": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "rpcStatus": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/protobufAny"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "v1LoginRequest": {
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "v1LoginResponse": {
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expTime": {
            "format": "int64",
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "v1LogoutRequest": {
        "properties": {
          "refresh": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "v1LogoutResponse": {
        "type": "object"
      },
      "v1Profile": {
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "v1RefreshRequest": {
        "properties": {
          "refresh": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "v1RegisterRequest": {
        "properties": {
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/v1Role"
          }
        },
        "type": "object"
      },
      "v1RegisterResponse": {
        "type": "object"
      },
      "v1Role": {
        "default": "ROLE_STAFF",
        "enum": [
          "ROLE_STAFF",
          "ROLE_ADMIN",
          "ROLE_DIRECTOR",
          "ROLE_MANAGER",
          "ROLE_SUPERVISOR"
        ],
        "type": "string"
      }
    }
  },
  "info": {
    "title": "auth.proto",
    "version": "version not set"
  },
  "openapi": "3.0.3",
  "paths": {
    "/v1/auth/login": {
      "post": {
        "operationId": "AuthService_Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/v1LoginRequest"
              }
            }
          },
          "required": true,
          "x-originalParamName": "body"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/v1LoginResponse"
                }
              }
            },
            "description": "A successful response."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/rpcStatus"
                }
              }
            },
            "description": "An unexpected error response."
          }
        },
        "tags": [
          "AuthService"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "AuthService_Logout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/v1LogoutRequest"
              }
            }
          },
          "required": true,
          "x-originalParamName": "body"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/v1LogoutResponse"
                }
              }
            },
            "description": "A successful response."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/rpcStatus"
                }
              }
            },
            "description": "An unexpected error response."
          }
        },
        "tags": [
          "AuthService"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "AuthService_Refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/v1RefreshRequest"
              }
            }
          },
          "required": true,
          "x-originalParamName": "body"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/v1LoginResponse"
                }
              }
            },
            "description": "A successful response."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/rpcStatus"
                }
              }
            },
            "description": "An unexpected error response."
          }
        },
        "tags": [
          "AuthService"
        ]
      }
    },
    "/v1/auth/register": {
      "post": {
        "operationId": "AuthService_Register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/v1RegisterRequest"
              }
            }
          },
          "required": true,
          "x-originalParamName": "body"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/v1RegisterResponse"
                }
              }
            },
            "description": "A successful response."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/rpcStatus"
                }
              }
            },
            "description": "An unexpected error response."
          }
        },
        "tags": [
          "AuthService"
        ]
      }
    },
    "/v1/me": {
      "get": {
        "operationId": "AuthService_Me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/v1Profile"
                }
              }
            },
            "description": "A successful response."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/rpcStatus"
                }
              }
            },
            "description": "An unexpected error response."
          }
        },
        "tags": [
          "AuthService"
        ]
      }
    }
  },
  "tags": [
    {
      "name": "AuthService"
    }
  ]
}